🎨 refractor modelci hub converter structure
